Started Working on admin pages Reorganized files API Rework After doing some research found that GET requests should not have request bodies. To make my code work with the standard modified all of my GET API requests to take url parameters instead and use Content-Type urlencoded
