+++ approot/__init__.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+SERVER_ROOT = Path(__file__).parent.parent
