+++ approot/data_managers/errors.py
@@ -1,0 +1,10 @@
+class InvalidActionError(Exception):
+    pass
+
+
+class NotFoundError(Exception):
+    pass
+
+
+class InvalidMimetypeError(Exception):
+    pass
