+++ webroot/templates/401.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>401 - Unauthorized</title>
+</head>
+<body>
+    <h1>401 - Unauthorized</h1>
+    <p>Sorry, you are not authorized to access this page!</p>
+    <p>Did you mean to show up here? <b>Funny Joke</b></p>
+    <div>
+        <a href="/">Go back to the homepage</a>
+    </div>
+</body>
+</html>
