+++ webroot/templates/index.html
@@ -226,7 +226,7 @@
         </div>
         <div class="col">
             <h4>My Account</h4>
-            <a href="#">Sign In</a>
+            <a href="{{ url_for('webpages.login') }}">Sign In</a>
             <a href="#">View Cart</a>
             <a href="#">My wishlist</a>
             <a href="#">Tracking Information</a>
