+++ webroot/templates/login.html
@@ -8,8 +8,10 @@
         const login_api_url = "{{ url_for('user_api.login') }}";
         const login_url = "{{ url_for('webpages.login') }}";
         const home_url = "{{ url_for('webpages.home') }}";
+        const error = "{{ error }}";
     </script>
     <script src="{{ url_for('static', filename='js/login/script.js') }}"></script>
+
 </head>
 <body>
     <h1>Flowers4U - Login</h1>
@@ -20,7 +22,11 @@
         <input type="password" name="password" id="password" required>
         <input id="sub_btn" type="submit" value="Login">
     </form>
+    <p id="error_msg">{{ error }}</p>
 
     <p>Don't have an account? <a href="{{ url_for('webpages.register') }}">Register</a></p>
+    <input type="hidden" name="redirect_url" value="{{ redirect_url }}"> <!-- Hidden input to store the redirect url from server -->
+
 </body>
+
 </html>
