+++ webroot/templates/register.html
@@ -32,5 +32,7 @@
         <input id="sub_btn" type="submit" value="Register">
     </form>
     <p>Already have an account? <a href="{{ url_for('webpages.login') }}">Login</a></p>
+
+    <input type="hidden" name="redirect_url" value="{{ redirect_url }}">
 </body>
 </html>
